Ajustes en unidades y combates
