+++ src/main/scala/model/Combat.scala
@@ -1,24 +1,26 @@
 package cl.uchile.dcc.citric
 package model
 
+import cl.uchile.dcc.citric.model.Units.PlayerCharacter
+
 import scala.math
 
-class Combat(attacker: PlayerCharacter, attacked: Units) {
+class Combat(attacker: PlayerCharacter, attacked: PlayerCharacter) {
   private val roll = attacker.rollDice()
-  private val AttackedHP= attacked.getHP
+  private val AttackedHP = attacked.getHP
   attacker.setAttack(roll)
 
   /*decision*/
-  private val damage: Int = defend(attacker, attacked)
+  private val damage: Int = defend(attacker, attacked: PlayerCharacter)
   /*private val damage: Int = evasion(attacker, attacked)*/
 
-  if(damage >= AttackedHP){
+  if (damage >= AttackedHP) {
     attacked.setHP(-AttackedHP)
   } else {
     attack(damage, attacked)
   }
 
-  private def attack(damage:Int, attacked: Units): Unit = {
+  private def attack(damage: Int, attacked: PlayerCharacter): Unit = {
     attacked.setHP(-damage)
   }
 
@@ -28,44 +30,15 @@ class Combat(attacker: PlayerCharacter, attacked: Units) {
     damage
   }
 
-  private def defend(attacker: PlayerCharacter, attacked: WildUnit): Int = {
-    val damage = math.max(1, roll + attacker.getAttack - attacked.getDefense)
-    damage
-  }
 
   private def evade(attacker: PlayerCharacter, attacked: PlayerCharacter): Unit = {
     val rollEva = attacked.rollDice()
-    if(rollEva + attacked.getEva > roll + attacker.getAttack) {
+    if (rollEva + attacked.getEva > roll + attacker.getAttack) {
       val damage = 0
     } else {
       val damage = attacker.getAttack
     }
   }
-
-
 }
-/*
-class Combat(attacker: PlayerCharacter, attacked: WildUnit) {
-  val roll: Int = attacker.rollDice()
-  private val AttackedHP = attacked.getHP
-  attacker.setAttack(roll)
 
-  /*decision*/
-  private val damage: Int = defend(attacker, attacked)
-  /*private val damage: Int = evasion(attacker, attacked)*/
 
-  if (damage >= AttackedHP) {
-    attacked.setHP(-AttackedHP)
-  } else {
-    attack(damage, attacked)
-  }
-
-  private def attack(damage: Int, attacked: WildUnit): Unit = {
-    attacked.setHP(-damage)
-  }
-
-  private def defend(attacker: PlayerCharacter, attacked: WildUnit): Int = {
-    val damage = math.max(1, roll + attacker.getAttack - attacked.defense)
-    damage
-  }
-}
