+++ src/main/scala/model/Panels/Tablero.scala
@@ -1,8 +1,8 @@
 package cl.uchile.dcc.citric
-package model
+package model.Panels
 
+import cl.uchile.dcc.citric.model.Units.PlayerCharacter
 import scala.collection.mutable.ArrayBuffer
-import scala.util.Random
 
 /** This class will organizate and inicialize the map of the game
  *
