+++ src/main/scala/model/Units/Units.scala
@@ -1,0 +1,10 @@
+package cl.uchile.dcc.citric
+package model.Units
+
+trait Units {
+  def getHP: Int
+  def setHP(increase: Int): Unit
+  def getDefense: Int
+  def setStars(increase: Int): Unit
+  def getAttack: Int
+}
