+++ src/main/scala/model/Units/WildUnit.scala
@@ -1,0 +1,5 @@
+package cl.uchile.dcc.citric
+package model.Units
+
+trait WildUnit extends Units {
+}
